[#175] Common namespace for gitlab, github
